requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
openai==1.3.5
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
whisper==1.1.10
torch==2.1.0
librosa==0.10.1
soundfile==0.12.1
gunicorn==21.2.0
eventlet==0.33.3